+++ client/src/components/boss/AddUser.jsx
@@ -55,7 +55,6 @@ const AddUser = () => {
         email: '',
         phoneNumber: '',
         address: '',
-        workingHours: '',
         dni: '',
         rol: [],
         file: '',
