+++ client/src/components/guard/GuardProfile.jsx
@@ -7,27 +7,11 @@ import {
   filterTaskByIdAndStatus,
   updateStatus,
 } from "../../redux/actions";
-import Modal from "../reusable/Modal";
 import EditState from "../EditState/EditState"; 
 import './styles.css'
 
 export default function GuardProfile() {
-  const [loading, setLoading] = useState(false);
-  const [image, setImage] = useState('')
-  const uploadImage = async (e) => {
-      const files = e.target.files;
-      const data = new FormData();
   
-      data.append('file', files[0]);
-      data.append('upload_preset', 'magqqp6o');
-     
-      setLoading(true);
-      const res = await fetch("https://api.cloudinary.com/v1_1/henrysecurityapp/image/upload", { method: "POST", body: data })
-      const file = await res.json();
-      
-      setImage(file.secure_url);
-      setLoading(false)
-  };
 
   const ToDos = useSelector((state) => state.todosId);
   const user = useSelector((state) => state.userDetails);
@@ -35,13 +19,8 @@ export default function GuardProfile() {
   const dispatch = useDispatch();
   const { id } = useParams();
 
-    const [active, setActive] = useState(false);
     const [currentState, setCurrentState] = useState("All")
 
-  const toggle = ()=>{
-    setActive(!active);
-    setImage('')
-  }
 
   useEffect(() => {
     dispatch(getUsersById(id));
@@ -106,58 +85,9 @@ export default function GuardProfile() {
           </div>
         ))}
         
-      </div>
-      <Modal active={active} toggle={toggle}>
-        <div style={style.modal}>
-          <label>Comentario:</label>
-          <textarea className={Input()} placeholder="Comentario sobre la tarea..."></textarea>
-            <input className={File} type="file" name="file" onChange={uploadImage}></input>
-            {loading ? (<p>...loading</p>) : (<img src={image} style={style.img}/>)}
-            <button onClick={()=>setImage('')}>x</button>
-          <button className={Button()}>Enviar</button>
         </div>
-      </Modal>
-    </div>
+      </div>
+      
+    
   );
 }
-
-const style = {
-  modal: {
-    display:'flex',
-    flexDirection: 'column',
-    margin: '5px'
-  },
-  img: {
-    width :'100px',
-    heigth: '100px'
-  }
-}
-
-const Input = (props) => `
-    hover:bg-slate-100
-    placeholder:italic placeholder:text-slate-400 
-    block bg-white w-${props === 'Select' ? '32' : '96'} m-2.5
-    border border-slate-300 rounded-md 
-    py-2 pl-9 pr-3 shadow-sm 
-    focus:outline-none focus:border-blue-500 focus:ring-blue-500 focus:ring-1 
-    sm:text-sm
-`;
-
-const Button = () => `
-    font-bold text-white
-    bg-blue-500
-    w-32 h-10 p-0 mt-3
-    border-2 border-blue-500
-    hover:border-blue-600 hover:bg-blue-600
-    active:border-blue-700 active:bg-blue-700
-    rounded-3xl
-`;
-
-const File = `
-    block w-full text-sm text-slate-500
-    file:mr-4 file:py-2 file:px-4
-    file:rounded-full file:border-0
-    file:text-sm file:font-semibold
-    file:bg-blue-50 file:text-blue-700
-    hover:file:bg-blue-100
-`;
